+++ spring-data-and-security-example/src/main/java/pl/marcinm312/springdatasecurityex/service/db/QuestionManager.java
@@ -5,8 +5,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import pl.marcinm312.springdatasecurityex.enums.Roles;
+import pl.marcinm312.springdatasecurityex.exception.ChangeNotAllowedException;
 import pl.marcinm312.springdatasecurityex.exception.ResourceNotFoundException;
 import pl.marcinm312.springdatasecurityex.model.Question;
+import pl.marcinm312.springdatasecurityex.model.User;
 import pl.marcinm312.springdatasecurityex.repository.QuestionRepository;
 
 @Service
@@ -29,23 +32,31 @@ public class QuestionManager {
 		}).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + questionId));
 	}
 
-	public Question createQuestion(Question question) {
+	public Question createQuestion(Question question, User user) {
+		question.setUser(user);
 		return questionRepository.save(question);
 	}
 
-	public Question updateQuestion(Long questionId, Question questionRequest) {
+	public Question updateQuestion(Long questionId, Question questionRequest, User user) {
 		return questionRepository.findById(questionId).map(question -> {
-			question.setTitle(questionRequest.getTitle());
-			question.setDescription(questionRequest.getDescription());
-			return questionRepository.save(question);
-
+			if (question.getUser().getId() == user.getId() || user.getRole().equals(Roles.ROLE_ADMIN.name())) {
+				question.setTitle(questionRequest.getTitle());
+				question.setDescription(questionRequest.getDescription());
+				return questionRepository.save(question);
+			} else {
+				throw new ChangeNotAllowedException("Change not allowed!");
+			}
 		}).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + questionId));
 	}
 
-	public boolean deleteQuestion(Long questionId) {
+	public boolean deleteQuestion(Long questionId, User user) {
 		return questionRepository.findById(questionId).map(question -> {
-			questionRepository.delete(question);
-			return true;
+			if (question.getUser().getId() == user.getId() || user.getRole().equals(Roles.ROLE_ADMIN.name())) {
+				questionRepository.delete(question);
+				return true;
+			} else {
+				throw new ChangeNotAllowedException("Change not allowed!");
+			}
 		}).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + questionId));
 	}
 }
